+++ finagle-redis/src/main/scala/com/twitter/finagle/redis/exp/SubscribeDispatcher.scala
@@ -2,13 +2,14 @@ package com.twitter.finagle.redis.exp
 
 import com.twitter.finagle.dispatch.GenSerialClientDispatcher
 import com.twitter.finagle.redis.protocol._
+import com.twitter.finagle.stats.StatsReceiver
 import com.twitter.finagle.transport.Transport
 import com.twitter.util.{Future, Promise, Return, Throw}
 import java.util.concurrent.atomic.AtomicReference
 import scala.util.control.NonFatal
 
-class SubscribeDispatcher(trans: Transport[Command, Reply])
-    extends GenSerialClientDispatcher[Command, Reply, Command, Reply](trans) {
+class SubscribeDispatcher(trans: Transport[Command, Reply], statsReceiver: StatsReceiver)
+    extends GenSerialClientDispatcher[Command, Reply, Command, Reply](trans, statsReceiver) {
 
   private val handler = new AtomicReference[SubscribeHandler]